Add bindings to change border-width per window
When I have windows with dark and light backgrounds next to each other,
it's tough to choose a border color that's visible unless the border is
quite thick. This isn't a common enough situation that I want to
increase the border width globally, though. Instead, with this change, I
can manually increase the border width for a particular situation until
it's visible enough.

An alternative solution I considered was to increase the border for
certain classes of windows when the `tile` layout was selected, but that
requires more code that's more likely to conflict with upstream updates.